requirements.txt
Django==3.1.5
django-debug-toolbar==3.2.1
django-simple-captcha==0.5.13

Pillow==8.1.2
psycopg2==2.8.6
